Reviewed logic and updated Main package
