+++ Client.xcodeproj/project.xcworkspace/xcshareddata/swiftpm/Package.resolved
@@ -50,7 +50,7 @@
       "location" : "https://github.com/ecosia/ios-core.git",
       "state" : {
         "branch" : "MOB-XXX_checkoff_whatsnewpage_user_profile",
-        "revision" : "b57db0c6c521e5aaeff26306b69d28a80ab5eca4"
+        "revision" : "e4176e5bde852a1d06bd8f32937436eb89cf887d"
       }
     },
     {
